+++ source/geh_common/release-notes.md
@@ -1,5 +1,11 @@
 # GEH Common Release Notes
 
+## Version 5.10.5
+
+- Adds `EnergySupplierIds` type for validating energy supplier IDs in pydantic models.
+- Refactors logic for `GridAreaCodes` as converters can be shared between different types.
+- Adds tests for `EnergySupplierIds` to ensure correct validation and conversion.
+
 ## Version 5.10.4
 
 - `physical_status` should be a string not a boolean